# LoggerMessage formatter performance improvements