@Scooletz 2021-09-03
@maryamariyan Thank you for chiming in. Looking forward for your opinion on whether this is something we want to pursue. This will require a lot of polishing, but the initial idea is in there. If you need anything, ping me.