@Scooletz 2021-09-21
It's been 20 days since a request to get some input. Could sb chime in and let me know whether this is something that has a possibility to be merged in? I don't want to propose sth to get it rejected.