@danmoseley 2021-09-21
It should not take this long.
@dotnet/area-extensions-logging @maryamariyan