@Scooletz 2021-09-23
Thx for chiming in @eerhardt . If this is looks like an initially interesting idea let me give it a spin and make CI green. Also, if there are some more opinions about it, let me know.