@stephentoub 2021-09-23
In https://github.com/dotnet/runtime/issues/50330 we discussed using the proposed CompositeFormat as an implementation detail under LoggerMessage.Define.  Is this essentially doing the same thing?