@Scooletz 2021-09-23
I wasn't aware of this beautiful proposal @stephentoub ! Actually, the case that I'm raising here, is mentioned in the original repo https://github.com/geeknoid/FastFormatting#template-based-format-strings ! Having this said, I'm closing this PR as the mentioned functionality is set to [Future milestone](https://github.com/dotnet/runtime/milestone/1). Logging will get this lunch for free 😅